add reply crud

+++ src/main/java/com/codestates/pre012/posts/service/PostsService.java
@@ -2,8 +2,11 @@ package com.codestates.pre012.posts.service;
 
 import com.codestates.pre012.exception.BusinessLogicException;
 import com.codestates.pre012.exception.ExceptionCode;
+import com.codestates.pre012.member.entity.Member;
 import com.codestates.pre012.posts.entity.Posts;
 import com.codestates.pre012.posts.repository.PostsRepository;
+import com.codestates.pre012.reply.ReplyRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -14,23 +17,25 @@ import java.util.Optional;
 
 @Service
 @Transactional
+@RequiredArgsConstructor
 public class PostsService {
 
     private final PostsRepository postsRepository;
 
-    public PostsService(PostsRepository postsRepository) {
-        this.postsRepository = postsRepository;
-    }
+    private final ReplyRepository replyRepository;
 
 
-    public Posts savedPosts(Posts postsPost) {
 
+    public Posts savedPosts(Posts postsPost, Member member) {
+
+        postsPost.setMember(member);
         return postsRepository.save(postsPost);
     }
 
-    public Posts updatePosts(Posts patchPost) {
+    public Posts updatePosts(Posts patchPost, Member member) {
 
         Posts findPosts = existPosts(patchPost.getPostsId());
+        if(!findPosts.getMember().equals(member)) throw new RuntimeException("자신의 글만 수정 가능합니다.");
 
         Optional.ofNullable(patchPost.getTitle())
                 .ifPresent(findPosts::setTitle);
@@ -52,11 +57,13 @@ public class PostsService {
         return postsRepository.findAll(PageRequest.of(page, size, Sort.by("postsId").descending()));
     }
 
-    public void deletePosts(long postId) {
+    public void deletePosts(long postId, Member member) {
 
         Posts findPosts = existPosts(postId);
-
-        postsRepository.delete(findPosts);
+        if(findPosts.getMember().equals(member)) {
+            postsRepository.delete(findPosts);
+        }
+        else throw new RuntimeException("자신의 게시글만 삭제 가능합니다.");
     }
 
     private Posts existPosts(long postsId) {
